requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
tomli==2.0.1
urllib3==1.26.8
validator==0.7.1
Werkzeug==2.0.3
